chore: split backend schemas.py into per-domain schemas/ package
